[techlag] Support extensibility
This code provides a mechanism to easily extend techlag to different
package managers (beyond Npm, the already supported one). In a nutshell,
every backend can be initialized with a combination of the following
parameters: `package`, `version` or `url`, and needs to implement
the `analyze` method (which is in charge of calculating the technical
lag.